+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2022-08-10
+### Added
+
+- Line Items (prices & cart details)
+- Klarna & Oney payment methods
+
+### Fixed
+
+- Cart checksum calculation now includes vouchers
+
 ## [1.2.1] - 2022-07-11
 ### Changed